no need for double

+++ clHCA.h
@@ -251,12 +251,12 @@ PACKED(
 
 	struct stChannel
 	{
-		double wave[8][0x80];
-		double base[0x80];
-		double block[0x80];
-		double wav1[0x80];
-		double wav2[0x80];
-		double wav3[0x80];
+		float wave[8][0x80];
+		float base[0x80];
+		float block[0x80];
+		float wav1[0x80];
+		float wav2[0x80];
+		float wav3[0x80];
 		uint8_t value[0x80];
 		uint8_t scale[0x80];
 		uint8_t value2[8];
